Create 09Taylor.c
a program in C to calculate the approximate value of exp (0.5) using the Taylor Series expansion for the exponential function. Use the terms in the expansion until the last term is less than the machine epsilon defines as FLT_EPSILON in the header file <float.h>.   Print the value returned by the Mathematical function exp ( ) also.