Create 13isPrimefn.c
a  function isprime(x) that  accepts   an   integer   argument  and   returns   1  if  the argument is prime and 0 otherwise. The function is to use plain division checking approach to determine if a given number is prime. Invoke this function from the main with different values obtained from the user and print appropriate messages.