V/6/USPCD/5blnCommand: Add code to display inode numbers
Currently after creation of the linked files we exit the program, any
sort of verification is done explicitly by the user. Streamline this
approach by adding code to display the inode numbers of created files
towards the end of the program, this provides an easier way to verify
the working of the program.

Signed-off-by: Karthik Nayak <Karthik.188@gmail.com>